fix : move index to root
